requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
astroid==3.2.2
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
concurrent-log-handler==0.9.25
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
Cython==0.29.37
dill==0.3.8
diskcache==5.6.3
Django==5.0.6
django-appconf==1.0.6
django-bulk-update==2.2.0
django-cors-headers==4.3.1
django-cryptography==1.1
django-extensions==3.2.3
django-filter==24.2
django-picklefield==3.2
django-q2==1.6.2
django-taggit==5.0.1
django4-chunked-upload==2.0.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dlib==19.24.4
drf-extensions==0.7.1
drf-yasg==1.21.7
face-recognition==1.3.0
face-recognition-models==0.3.0
faiss-cpu==1.8.0
filelock==3.14.0
Flask==3.0.3
Flask-RESTful==0.3.10
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.6.0
geographiclib==2.0
geopy==2.4.1
gevent==24.2.1
greenlet==3.0.3
gunicorn==22.0.0
h3==3.7.7
hdbscan==0.8.36
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
inflection==0.5.1
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
llama_cpp_python==0.2.78
MarkupSafe==2.1.5
matplotlib==3.9.0
mccabe==0.7.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
onnxruntime==1.18.0
packaging==24.1
pandas==2.2.2
pathvalidate==3.2.0
pillow==10.3.0
platformdirs==4.2.2
portalocker==2.8.2
protobuf==5.27.1
psycopg==3.1.19
pycparser==2.22
PyExifTool==0.4.9
PyJWT==2.8.0
pylint==3.2.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
pyvips==2.2.3
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
sqlparse==0.5.0
sympy==1.12.1
threadpoolctl==3.5.0
timezonefinder==6.5.0
timm==1.0.3
tokenizers==0.19.1
tomlkit==0.12.5
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.41.2
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
Wand==0.6.13
Werkzeug==3.0.3
zope.event==5.0
zope.interface==6.4.post2
